refactor(client): impl withRetrieveToken apiRoot extension based on retrieveToken function
